feat: remove Google Analytics and the cookie banner
Remove the support for project specific Google analytics codes and
replace it with Matomo.

We need to change Docs Italia's code in order to have feature parity
and supporting project specific Matomo siteIds, so for now the
id is empty for everyone until we have that feature.

Also remove the cookie banner because we are not serving tracking
cookies.